Refactoring
Moved login and registration functionality to the login and registration pages respectively, AppId now a field in the Gradle File
