+++ ACE_Application/app/src/main/java/com/ace_club_application/app/MainActivity.java
@@ -18,16 +18,16 @@ import android.widget.EditText;
 import android.widget.TextView;
 
 public class MainActivity extends AppCompatActivity {
-    String appId = "ace_android_application-sqneg";
-    // ACE APP ID = "ace_android_application-sqneg"
-
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         //Displays main page and initializes MongoDB realm on application
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        //Initializes MongoDB realm on application
+        Realm.init(this);
+        App app = new App(new AppConfiguration.Builder(BuildConfig.REALM_APP_ID).build());
 
         Intent intent = getIntent();
         Bundle extras = intent.getExtras();
@@ -54,42 +54,38 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
-        //Initializes MongoDB realm on application
-        Realm.init(this);
-        App app = new App(new AppConfiguration.Builder(appId).build());
-
         //If we have a valid email and password
         if (email != null && password != null)
         {
             //If we're coming from the login page, login to the app
             if (login)
             {
-                Credentials credentials = Credentials.emailPassword(email, password);
-                app.loginAsync(credentials, new App.Callback<User>() {
-                    @Override
-                    public void onResult(App.Result<User> result) {
-                        if (result.isSuccess())
-                        {
-                            Log.v("User", "logged in via email and password");
-                        }
-                        else
-                        {
-                            Log.v("User", "failed to login");
-                        }
-                    }
-                });
+//                Credentials credentials = Credentials.emailPassword(email, password);
+//                app.loginAsync(credentials, new App.Callback<User>() {
+//                    @Override
+//                    public void onResult(App.Result<User> result) {
+//                        if (result.isSuccess())
+//                        {
+//                            Log.v("User", "logged in via email and password");
+//                        }
+//                        else
+//                        {
+//                            Log.v("User", "failed to login");
+//                        }
+//                    }
+//                });
             }
             //Otherwise, create a new user with the email and password.
             else
             {
-                app.getEmailPassword().registerUserAsync(email, password, it->{
-                    if (it.isSuccess()){
-                        Log.v("User", "registered with email");
-                    }
-                    else{
-                        Log.v("User", "was unsuccessful in registration");
-                    }
-                });
+//                app.getEmailPassword().registerUserAsync(email, password, it->{
+//                    if (it.isSuccess()){
+//                        Log.v("User", "registered with email");
+//                    }
+//                    else{
+//                        Log.v("User", "was unsuccessful in registration");
+//                    }
+//                });
             }
         }
